Added a Carbon Copy feature in the email. Made it so the Email gets the customer data. Made it so the Email sends the wall data. Removed unnecessary imports from the class. Refactor the code so the class Email.java is located in the model Directory.
