+++ 04 Implementation/app/src/main/java/com/example/newyorker/model/Email.java
@@ -1,14 +1,7 @@
-package com.example.newyorker.view;
-import android.content.Intent;
-import android.net.Uri;
-import android.util.Log;
-
-import com.example.newyorker.model.Customer;
-import com.example.newyorker.model.Wall;
+package com.example.newyorker.model;
 
 import java.io.Serializable;
 import java.util.concurrent.ExecutionException;
-
 import uk.co.jakebreen.sendgridandroid.SendGrid;
 import uk.co.jakebreen.sendgridandroid.SendGridMail;
 import uk.co.jakebreen.sendgridandroid.SendGridResponse;
@@ -16,16 +9,11 @@ import uk.co.jakebreen.sendgridandroid.SendTask;
 
 
 public class Email implements Serializable {
-    private final String toNewYorker = "fredrikbille@hotmail.com";
-
-    private final String subject = "Forespørgsel New Yorker væg";
+    private final String toNewYorker = "hjordrup96@live.dk";
 
 
-    public void sendEmail(){
+    public void sendEmail(Customer customer, Wall wall){
 
-
-        /*String[] to = {forhandler};
-        String[] cc = {toNewYorker};
 
         String body = "Kunde navn: " + customer.getCustomerName() +
                 "\nKunde E-mail: " + customer.getCustomerEmailAddress() +
@@ -36,27 +24,26 @@ public class Email implements Serializable {
                 "\nVæg Højde: " + wall.getWallHeight() + " Cm." +
                 "\nVæg Pris: " + Math.round(wall.getWallPrice() * 100.0) / 100.0 + " Kr." +
                 "\n\n" + "Kunde Noter: " +
-                "\n"+customer.getCustomerNotes();*/
+                "\n"+customer.getCustomerNotes();
 
-
-        Log.d("András","Mailen oprettes.");
         SendGrid sendGrid = SendGrid.create("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");
         SendGridMail mail = new SendGridMail();
         mail.addRecipient("fredrikbille@hotmail.com", "Bitch");
-        mail.setFrom("NYWallBuilder@gmail.com", "Bille");
-        mail.setSubject("Hej fra yeppers");
-        mail.setContent("Hej, det virker nu. DO YOU EVEN SEND!?.");
-        Log.d("András","Mailen er nu konstrueret.");
-        Log.d("András","Mailen sendes.");
+        mail.addRecipientCarbonCopy(toNewYorker, "NewYorker");
+        mail.setFrom("NYWallBuilder@gmail.com", "NewYorkerApp");
+        mail.setSubject("Forespørgsel New Yorker væg");
+        mail.setContent(body);
+
         SendTask task = new SendTask(sendGrid, mail);
-        try {
+
+
+        /*try {
             SendGridResponse response = task.execute().get();
-            Log.d("András", "Mailen er sendt til APIet" + response.getErrorMessage() );
         } catch (ExecutionException e) {
             e.printStackTrace();
         } catch (InterruptedException e) {
             e.printStackTrace();
-        }
+        }*/
     }
